Move ping into daemon.
