+++ daemon/src/main.rs
@@ -1,6 +1,7 @@
+use daemon::process;
 use std::{error::Error, net::SocketAddrV4};
 use structopt::StructOpt;
-use tokio::net::{TcpListener, TcpStream};
+use tokio::net::TcpListener;
 
 #[derive(StructOpt)]
 #[structopt(name = "7roxy Daemon", about = "A personal AI proxy agent daemon.")]
@@ -9,16 +10,11 @@ struct Opt {
     local_address: SocketAddrV4,
 }
 
-async fn process(socket: TcpStream) {
-}
-
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
-    let Opt {
-        local_address,
-    } = Opt::from_args();
+    let Opt { local_address } = Opt::from_args();
 
-    let mut listener = TcpListener::bind(local_address).await?;
+    let mut listener = TcpListener::bind(&local_address).await?;
 
     loop {
         let (socket, _) = listener.accept().await?;
